Normalize text renderer crate root
